git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@3954 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwEUtils/Core/ITaxonData.vb
@@ -47,7 +47,7 @@ Namespace Core
         ''' <summary>Western limit of the bounding box where this taxon occurs.</summary>
         Property West() As Single
         ''' <summary>Julian date when this taxon was last updated.</summary>
-        Property LastUpdated() As Single
+        Property LastUpdated() As Double
 
     End Interface
